refactor(sidebar): extract ProjectRow, ChildRow + inline-edit/drag hooks from ProjectsSection
Resolves #414

ProjectsSection.tsx was 1,645 lines; it now weighs 737 lines (55% smaller).
The four extracted units each live in a dedicated sibling file:

- ProjectRow.tsx (193 lines) — project folder row component
- ChildRow.tsx (220 lines) — expanded child entry row component
- useProjectInlineEdit.ts (297 lines) — rename/create signal handling hook
- useProjectRowDrag.ts (24 lines) — file drag start logic hook
- project-section-utils.ts (140 lines) — shared pure utilities + RowDescriptor

Backward-compat re-exports (countMarkdownFiles, isSystemFolderName,
buildProjectNameValidator, RowDescriptor) keep existing tests green without
any import-path changes in ProjectsSection.test.tsx.
